perf(sync-server): probe the array branch with one && per batch (#9503)
The batch array branch probed the entity_ids GIN once per requested id.
Measured by A/B-ing the real statement with only that CTE swapped, on
postgres:16-alpine 16.14 at 1.5M rows with a 100-id all-new probe:

  fastupdate            @> per id            &&
  off (production)       12.4 ms /    700 blk   0.64 ms /   601 blk
  on  (prisma db push)    261 ms / 47,800 blk    170 ms / 1,072 blk

Two independent causes. A generic plan cannot see the 100 bound ids, so
re-probing per opaque parameter costs ~18x one multi-key probe. And a GIN
scan reads the whole fastupdate pending list, so N probes read it N times.
Both callers chunk at 100, so a 1000-id op reached 2.6 s inside the upload
transaction on a db-push database — #9503's own failure mode, re-created by
its fix. Production has fastupdate = off and was not exposed; CI, the E2E
stack and the documented manual setup are.

The earlier per-id form was justified by PGlite/PG18 abandoning the GIN for
a seq scan under a dirty pending list. The plan claim holds; the cost
conclusion was backwards — that seq scan is 6.8x faster and reads 20x fewer
blocks than the GIN plan @> gets. The dirty-pending-list test asserted
not.toContain('Seq Scan'), pinning the worse outcome as the requirement; it
now asserts cost, plus a canary that the pending list is really dirty.

Equivalence verified by differential fuzzing: 448 comparisons over
randomised ops and probes, zero differences, with controls that drop the
INTERSECT or weaken && to @> diverging in 220/224 and 173/224 rounds.

Also fixes a real defect in conflict-entity-lookup-plan.pglite.spec.ts: its
copy of the plan walker summed per-loop averages as totals and ignored
Rows Removed by Join Filter, so a join fan-out scored 0. The harness is now
shared with the batch spec (explain-plan.helper.ts) with the fix in one
place, which is what the divergence argued for.

Comment corrections, all measured against the real statement rather than a
hand-written approximation: the scale trade for small accounts (0.05 ms ->
0.59 ms, not the previously claimed inversion), the row-touch ceilings, and
the ::text cast rationale (inference resolves them; they are belt and
braces, not a requirement).
